+++ src/system/ui/UIEnums.h
@@ -1,0 +1,24 @@
+#ifndef UI_UIENUMS_H
+#define UI_UIENUMS_H
+
+enum UIListWidgetState {
+    kUIListWidgetActive,
+    kUIListWidgetHighlight,
+    kUIListWidgetInactive,
+    kNumUIListWidgetStates
+};
+
+enum UIListWidgetDrawType {
+    kUIListWidgetDrawAlways,
+    kUIListWidgetDrawOnlyFocused,
+    kUIListWidgetDrawNever,
+    kNumUIListWidgetDrawTypes
+};
+
+enum DrawCommand {
+    kDrawAll,
+    kDrawFirst,
+    kExcludeFirst
+};
+
+#endif
