+++ src/system/ui/UIListWidget.h
@@ -4,24 +4,13 @@
 #include "ui/UIColor.h"
 #include "math/Vec.h"
 #include "ui/UIComponent.h"
+#include "ui/UIListState.h"
+#include "ui/UIListProvider.h"
+#include "ui/UIEnums.h"
 #include "obj/ObjPtr_p.h"
 #include <vector>
 
 class UIList;
-
-enum UIListWidgetState {
-    kUIListWidgetActive,
-    kUIListWidgetHighlight,
-    kUIListWidgetInactive,
-    kNumUIListWidgetStates
-};
-
-enum UIListWidgetDrawType {
-    kUIListWidgetDrawAlways,
-    kUIListWidgetDrawOnlyFocused,
-    kUIListWidgetDrawNever,
-    kNumUIListWidgetDrawTypes
-};
 
 class UIListWidgetDrawState {
 public:
@@ -47,11 +36,11 @@ public:
     virtual int SubList(int){ return 0; }
     virtual void ResourceCopy(const UIListWidget*);
     virtual void CreateElements(UIList*, int){}
-    // virtual void Draw(const UIListWidgetDrawState&, const UIListState&, const Transform&, UIComponent::State, Box*, DrawCommand){}
-    // virtual void Fill(const UIListProvider&, int, int, int){}
-    // virtual void StartScroll(int, bool){}
-    // virtual void CompleteScroll(const UIListState&, int){}
-    // virtual void Poll(){}
+    virtual void Draw(const UIListWidgetDrawState&, const UIListState&, const Transform&, UIComponent::State, Box*, DrawCommand){}
+    virtual void Fill(const UIListProvider&, int, int, int){}
+    virtual void StartScroll(int, bool){}
+    virtual void CompleteScroll(const UIListState&, int){}
+    virtual void Poll(){}
 
     float DrawOrder() const;
     float DisabledAlphaScale() const;
